Move insertValue method to server
